Type safe action dispatching through ActionType type class.

+++ diode-core/shared/src/main/scala/diode/util/RunAfter.scala
@@ -13,7 +13,7 @@ trait RunAfter {
   def runAfter[A](delay: FiniteDuration)(f: => A): Future[A]
 
   def runAfter[A](delay: Int)(f: => A): Future[A] =
-    runAfter(FiniteDuration(delay, TimeUnit.MILLISECONDS))(f)
+    runAfter(FiniteDuration(delay.toLong, TimeUnit.MILLISECONDS))(f)
 }
 
 object RunAfter {
